// client/projects/console/src/app/communities/homescreen/models/HomeScreenBottomNavigation.ts
// tslint:disable
/**
 * Our City App
 * Our City App internal apis
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import {
    HomeScreenNavigationButton,
} from './';

/**
 * Content of the bottom navigation bar
 * @export
 * @interface HomeScreenBottomNavigation
 */
export interface HomeScreenBottomNavigation {
    /**
     * @type {Array<HomeScreenNavigationButton>}
     * @memberof HomeScreenBottomNavigation
     */
    buttons?: Array<HomeScreenNavigationButton>;
}