// client/projects/console/src/app/communities/homescreen/models/HomeScreenNavigationButton.ts
// tslint:disable
/**
 * Our City App
 * Our City App internal apis
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * A button shown in the bottom navigation of the app
 * @export
 * @interface HomeScreenNavigationButton
 */
export interface HomeScreenNavigationButton {
    /**
     * @type {string}
     * @memberof HomeScreenNavigationButton
     */
    icon: string;
    /**
     * @type {string}
     * @memberof HomeScreenNavigationButton
     */
    label: string;
    /**
     * @type {string}
     * @memberof HomeScreenNavigationButton
     */
    action: string;
}